+++ sphinx/source/changelog.rst
@@ -11,9 +11,10 @@ Changelog
 -----
 * Added the :ref:`tap2sna.py` command (for building snapshots from TAP/TZX
   files)
+* Moved the ``StyleSheet`` parameter from the ``[Paths]`` section to the
+  :ref:`ref-game` section
 * Moved the ``JavaScript`` parameter from the ``[Paths]`` section to the
   :ref:`page` section
-
 
 3.4 (2013-07-08)
 ----------------